refactor(util): Move `backslashEscapeString` to `StringUtil` class
