+++ src/main/java/de/rub/nds/modifiablevariable/string/ModifiableString.java
@@ -7,7 +7,7 @@
  */
 package de.rub.nds.modifiablevariable.string;
 
-import static de.rub.nds.modifiablevariable.util.ArrayConverter.backslashEscapeString;
+import static de.rub.nds.modifiablevariable.util.StringUtil.backslashEscapeString;
 
 import de.rub.nds.modifiablevariable.ModifiableVariable;
 import de.rub.nds.modifiablevariable.VariableModification;
